drm: msm: dsi_ctrl: Set failed clock to NULL to avoid KP on deinit
When we fail probing some clock we go deinit whatever was
previously initialized and, to do this, we are checking for
value in the clock related variable.
Problem is, devm_clk_get returns a ERR_PTR in case of failure
and we definitely don't want to unregister something that
did never register, or we will end up with a not-so-nice kernel
panic.

To address this, set the pointers to NULL whenever a devm_clk_get
call fails for any clock after saving the error for later.